Split source move rollback logic
